+++ @xen-orchestra/lite/CHANGELOG.md
@@ -1,5 +1,9 @@
 # ChangeLog
 
+## **next**
+
+- [XOA deploy] Fix the error and success screens never being displayed at the end of a deployment (PR [#10232](https://github.com/vatesfr/xen-orchestra/pull/10232))
+
 ## **0.24.0** (2026-07-30)
 
 - [Host/dashboard] Switch CPU and RAM panels order to match Pool dashboard layout (PR [#10059](https://github.com/vatesfr/xen-orchestra/pull/10059))